Logo and splash screen

+++ Assets/Scripts/Simulation/Models/Stat.cs
@@ -4,9 +4,9 @@ using System.Collections;
 public class Stat {
 
   public StatType Type { get; set; }
-  public int Value { get; set; }
+  public float Value { get; set; }
 
-  public Stat (string statKey, int value) {
+  public Stat (string statKey, float value) {
     Type = StatType.all[statKey];
     Value = value;
   }
